Add CurlParamProps component
