+++ components/layout/execute/index.ts
@@ -1,5 +1,6 @@
 export * from "./CurlBodyProps";
 export * from "./CurlCommand";
+export * from "./CurlParamProps";
 export * from "./ExecutePanel";
 export * from "./ExecuteResponseExample";
 export * from "./ExecuteHeader";
